refactor(auth): Update AuthService to implement an interface
