+++ backend/src/main/java/com/project/G1_T3/authentication/controller/AuthenticationController.java
@@ -1,6 +1,6 @@
 package com.project.G1_T3.authentication.controller;
 
-import com.project.G1_T3.authentication.service.AuthService;
+import com.project.G1_T3.authentication.service.AuthServiceImpl;
 import com.project.G1_T3.user.model.UserDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.*;
 public class AuthenticationController {
 
     @Autowired
-    private AuthService authService;
+    private AuthServiceImpl authService;
 
     @PostMapping("/validate-token")
     public ResponseEntity<UserDTO> validateToken(@RequestHeader("Authorization") String token) {
